Tweak Toyota Tuning to Improve Acceleration from a Stop
The Rav4 2019 TSS2 has ~1.4s lag between when accel commands are sent
to the car and when they are acted upon when the car is starting from
a dead stop.  For freeway driving traffic, the accel from a dead stop
is slow enough to make this barely noticable, which is what Toyota was
developing for.

However, for street driving, for example at a stop light, the 1.4s lag
is very noticable and somewhat problematic.

The only solution, is to rapidly accellerate after the 1.4s lag to
catch up.  This of course causes a decent amount of jerk, so only so
much can be done before this become uncomfortable.

Changes:

Decrease minSpeedCan, vEgoStopping, vEgoStarting to 0.2.  Car is perfectly
capable of controlling at these speeds.  Lower value allows for state
machine to enter starting phase just a bit faster.

Set starting accel to 0.0 and the rate very high. This puts the car into
the ready to go phase as fast as possible.

Increase the ActuatorDelay lag to .3 so we look into the future just a bit
more.

Tweak stop accel and rate, the OEM stop_accel is only -.4.  This doesn't
have much to do with the sluggish start, just making things a bit smoother

Increase the PID max accel to 1.6 from 1.5. * Could probably try increasing
this slightly further.

Increase planner max accel from a stop when behind a lead to 1.6.

Slight tweaks in longcontrol to make sure stopAccel and startAccel are
properly respected.